+++ run.py
@@ -12,7 +12,7 @@ def main(settings):
     """Run the programm."""
     args = settings.args
     expenses = Expenses(
-        args.file, args.full_name, args.months, args.yearly, args.income, args.expense
+        args.file, args.months, args.yearly, args.income, args.expense
     )
     expenses.show()